remove unnecessary int cast in line 432

+++ src/pl/enhancer_plot.py
@@ -429,7 +429,7 @@ def enhancer_plot(
     # Use numpy's PTP (Peak-To-Peak) function to calculate range_size
     range_size = np.ptp(np.append(top_df["logFC"].values, 0)).astype(int)
 
-    ypaddings = np.linspace(0, int(max_len * range_size * point_distance), num=max_len)
+    ypaddings = np.linspace(0, max_len * range_size * point_distance, num=max_len)
 
 
     fontsize = _calculate_fontsize(ax, n_genes)
